refactor: add SearchResults
