+++ src/app/Scenes/Search/Search.tsx
@@ -15,13 +15,7 @@ import { useAlgoliaIndices } from "app/utils/useAlgoliaIndices"
 import { useSearchInsightsConfig } from "app/utils/useSearchInsightsConfig"
 import { Box, Flex, Spacer, Text, Touchable } from "palette"
 import { Suspense, useCallback, useEffect, useMemo, useRef, useState } from "react"
-import {
-  Configure,
-  connectInfiniteHits,
-  connectSearchBox,
-  connectStateResults,
-  InstantSearch,
-} from "react-instantsearch-native"
+import { Configure, connectSearchBox, InstantSearch } from "react-instantsearch-native"
 import { Keyboard, Platform, ScrollView } from "react-native"
 import {
   FetchPolicy,
@@ -32,8 +26,6 @@ import {
 } from "react-relay"
 import { useTracking } from "react-tracking"
 import styled from "styled-components"
-import { AutosuggestResult, AutosuggestResults } from "./AutosuggestResults"
-import { AlgoliaSearchResults } from "./components/AlgoliaSearchResults"
 import { CityGuideCTA } from "./components/CityGuideCTA"
 import { CityGuideCTANew } from "./components/CityGuideCTANew"
 import { SearchPlaceholder } from "./components/placeholders/SearchPlaceholder"
@@ -43,32 +35,16 @@ import { ALLOWED_ALGOLIA_KEYS, DEFAULT_PILLS, TOP_PILL } from "./constants"
 import { getContextModuleByPillName, isAlgoliaApiKeyExpiredError } from "./helpers"
 import { RecentSearches } from "./RecentSearches"
 import { RefetchWhenApiKeyExpiredContainer } from "./RefetchWhenApiKeyExpired"
-import { SearchArtworksQueryRenderer } from "./SearchArtworksContainer"
 import { SearchContext, useSearchProviderValues } from "./SearchContext"
+import { SearchResults } from "./SearchResults"
 import { AlgoliaIndexKey } from "./types"
-import { AlgoliaSearchResult, PillType } from "./types"
-
-interface TappedSearchResultData {
-  query: string
-  type: string
-  position: number
-  contextModule: ContextModule
-  slug: string
-  objectTab?: string
-}
+import { PillType } from "./types"
 
 const SearchInputContainer = connectSearchBox(SearchInput)
-const SearchResultsContainerWithState = connectStateResults(AlgoliaSearchResults)
-const SearchResultsContainer = connectInfiniteHits(SearchResultsContainerWithState)
 
 interface SearchState {
   query?: string
   page?: number
-}
-
-const objectTabByContextModule: Partial<Record<ContextModule, string>> = {
-  [ContextModule.auctionTab]: "Auction Results",
-  [ContextModule.artistsTab]: "Artworks",
 }
 
 interface RefreshQueryOptions {
@@ -191,30 +167,6 @@ export const Search: React.FC = () => {
     setSearchState((prevState) => ({ ...prevState }))
   }
 
-  const renderResults = () => {
-    if (selectedPill.type === "algolia") {
-      return (
-        <SearchResultsContainer
-          selectedPill={selectedPill}
-          onRetry={handleRetry}
-          trackResultPress={handleTrackAlgoliaResultPress}
-        />
-      )
-    }
-    if (selectedPill.key === TOP_PILL.key) {
-      return (
-        <AutosuggestResults
-          query={searchState.query!}
-          showResultType
-          showQuickNavigationButtons
-          showOnRetryErrorMessage
-          trackResultPress={handleTrackAutosuggestResultPress}
-        />
-      )
-    }
-    return <SearchArtworksQueryRenderer keyword={searchState.query!} />
-  }
-
   const handlePillPress = (pill: PillType) => {
     const contextModule = getContextModuleByPillName(selectedPill.displayName)
 
@@ -231,36 +183,6 @@ export const Search: React.FC = () => {
   const handleResetSearchInput = () => {
     searchPillsRef?.current?.scrollTo({ x: 0, y: 0, animated: true })
     setSelectedPill(TOP_PILL)
-  }
-
-  const handleTrackAutosuggestResultPress = (result: AutosuggestResult, itemIndex?: number) => {
-    trackEvent(
-      tracks.tappedSearchResult({
-        type: result.displayType || result.__typename,
-        slug: result.slug!,
-        position: itemIndex!,
-        query: searchState.query!,
-        contextModule: ContextModule.topTab,
-      })
-    )
-  }
-
-  const handleTrackAlgoliaResultPress = (result: AlgoliaSearchResult) => {
-    const contextModule = getContextModuleByPillName(selectedPill.displayName)
-
-    const data: TappedSearchResultData = {
-      type: selectedPill.displayName,
-      slug: result.slug,
-      position: result.__position - 1,
-      query: searchState.query!,
-      contextModule: contextModule!,
-    }
-
-    if (contextModule && objectTabByContextModule[contextModule]) {
-      data.objectTab = objectTabByContextModule[contextModule]
-    }
-
-    trackEvent(tracks.tappedSearchResult(data))
   }
 
   return (
@@ -298,7 +220,11 @@ export const Search: React.FC = () => {
                     isSelected={isSelected}
                   />
                 </Box>
-                {renderResults()}
+                <SearchResults
+                  selectedPill={selectedPill}
+                  query={searchQuery}
+                  onRentry={handleRetry}
+                />
               </>
             ) : (
               <Scrollable>
@@ -387,16 +313,6 @@ const tracks = {
     subject,
     query,
     action: ActionType.tappedNavigationTab,
-  }),
-  tappedSearchResult: (data: TappedSearchResultData) => ({
-    context_screen_owner_type: Schema.OwnerEntityTypes.Search,
-    context_screen: Schema.PageNames.Search,
-    query: data.query,
-    position: data.position,
-    selected_object_type: data.type,
-    selected_object_slug: data.slug,
-    context_module: data.contextModule,
-    action: Schema.ActionNames.SelectedResultFromSearchScreen,
   }),
 }
